+++ pybamm/models/submodels/particle/fickian/fickian_many_particles.py
@@ -46,7 +46,29 @@ class ManyParticles(BaseModel):
         N_s = self._flux_law(c_s, T_k)
 
         variables.update(self._get_standard_flux_variables(N_s, N_s))
+
+        if self.domain == "Negative":
+            x = pybamm.standard_spatial_vars.x_n
+            R = pybamm.FunctionParameter("Negative particle distribution in x", x)
+            variables.update({"Negative particle distribution in x": R})
+
+        elif self.domain == "Positive":
+            x = pybamm.standard_spatial_vars.x_p
+            R = pybamm.FunctionParameter("Positive particle distribution in x", x)
+            variables.update({"Positive particle distribution in x": R})
         return variables
+
+    def set_rhs(self, variables):
+
+        c, N, _ = self._unpack(variables)
+
+        if self.domain == "Negative":
+            R = variables["Negative particle distribution in x"]
+            self.rhs = {c: -(1 / (R ** 2 * self.param.C_n)) * pybamm.div(N)}
+
+        elif self.domain == "Positive":
+            R = variables["Positive particle distribution in x"]
+            self.rhs = {c: -(1 / (R ** 2 * self.param.C_p)) * pybamm.div(N)}
 
     def _unpack(self, variables):
         c_s = variables[self.domain + " particle concentration"]